// packages/compiler/src/generated-property-specs.ts
export interface GeneratedSpec {
  readonly property: string;
  readonly syntax?: string;
  readonly initialValue?: string;
  readonly animatable: boolean;
  readonly inherits: boolean;
}